engajar/notificacao_indicadores/requirements.txt
Flask==2.3.2
gunicorn==20.1.0
firebase-admin==5.2.0
google-cloud-storage==2.3.0
google-cloud-firestore==2.7.2
Werkzeug==2.3.3
